requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astunparse==1.6.3
autograd==1.7.0
autoray==0.7.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diastatic-malt==2.15.2
exceptiongroup==1.2.2
fastapi==0.115.12
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.4.2
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
PennyLane==0.41.0
PennyLane_Lightning==0.41.0
protobuf==5.29.4
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
rustworkx==0.16.0
scikit-learn==1.6.1
scipy==1.15.2
scipy-openblas32==0.3.29.0.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
threadpoolctl==3.6.0
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
Werkzeug==3.1.3
wrapt==1.17.2
